+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Extension functions creating `RoomDatabase.Builder`:
   - `Context.roomDatabaseBuilder()`
   - `Context.roomInMemoryDatabaseBuilder()`
+- Annotation processor generating `TypeConverter`s for enum classes